First Playtest of v0.2 & Bugs Discovered After the first playtest there were much improvements, as well as some of the bugs being rather resilient.
These bugs have to do with the borders and them not being equal. Along with the screen flickering on the home page. Also the input keeps reloading quickly.

Need to see what methods are being used and if the engine is the problem. Overall much progress and looking good.

bye

signed agadapada